Move checkAdmin to commons.py
