+++ transmission.py
@@ -3,7 +3,7 @@ import yaml
 from telegram import ReplyKeyboardMarkup
 from telegram.ext import ConversationHandler
 
-from commons import checkId, authentication
+from commons import checkId, authentication, checkAdmin
 from definitions import CONFIG_PATH, LANG_PATH, ADMIN_PATH
 
 config = yaml.safe_load(open(CONFIG_PATH, encoding="utf8"))
@@ -99,19 +99,3 @@ def changeSpeedTransmission(update, context):
                 text=transcript["Transmission"]["ChangedToNormal"],
             )
             return ConversationHandler.END
-
-# Check if user is an admin
-def checkAdmin(update):
-    admin = False
-    user = update.message.from_user
-    with open(ADMIN_PATH, "r") as file:
-        for line in file:
-            if line.strip("\n") == str(user["username"]) or line.strip("\n") == str(
-                user["id"]
-            ):
-                admin = True
-        file.close()
-        if admin:
-            return True
-        else:
-            return False
